+++ backend/srs/analysis/DataProcessor.hpp
@@ -9,8 +9,8 @@
 #include <spdlog/logger.h>
 #include <srs/analysis/DataProcessManager.hpp>
 #include <srs/data/DataStructs.hpp>
-#include <srs/serializers/SerializableBuffer.hpp>
-#include <srs/serializers/StructDeserializer.hpp>
+#include <srs/converters/SerializableBuffer.hpp>
+#include <srs/converters/StructDeserializer.hpp>
 #include <srs/utils/CommonDefitions.hpp>
 #include <srs/writers/DataWriter.hpp>
 #include <tbb/concurrent_queue.h>
